+++ CHANGELOG.md
@@ -2,3 +2,4 @@ Changelog
 
 alpha1.3.1
 - first version
+- added AddExtract action to make use of the Extractor for custom items